requirements.txt
# Dollar Bill Serial Number Processor
# Python 3.8+

# Core dependencies
# Use headless OpenCV to avoid Qt conflicts with PySide6
opencv-python-headless>=4.8.0
numpy>=1.24.0
easyocr>=1.7.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Configuration
Pillow>=10.0.0
pyyaml>=6.0

# GUI (Phase 3)
PySide6>=6.5.0

# Export (Phase 4)
pandas>=2.0.0
openpyxl>=3.1.0
